+++ R/copy.R
@@ -76,14 +76,14 @@ dir_copy <- function(path, new_path) {
     }
     dir_create(new_path[[i]])
 
-    dirs <- dir_ls(path[[i]], type = "directory", recursive = TRUE, all = TRUE)
+    dirs <- dir_ls(path[[i]], type = "directory", recurse = TRUE, all = TRUE)
     dir_create(path(new_path[[i]], path_rel(dirs, path[[i]])))
 
-    files <- dir_ls(path, recursive = TRUE,
+    files <- dir_ls(path, recurse = TRUE,
       type = c("unknown", "file", "FIFO", "socket", "character_device", "block_device"), all = TRUE)
     file_copy(files, path(new_path[[i]], path_rel(files, path[[i]])))
 
-    links <- dir_ls(path, recursive = TRUE,
+    links <- dir_ls(path, recurse = TRUE,
       type = "symlink", all = TRUE)
     link_copy(links, path(new_path[[i]], path_rel(links, path[[i]])))
   }