+++ R/list.R
@@ -14,7 +14,8 @@
 #'
 #' @param type File type(s) to return, one or more of "any", "file", "directory",
 #'   "symlink", "FIFO", "socket", "character_device" or "block_device".
-#' @param recursive Should directories be listed recursively?
+#' @param recurse If `TRUE` recurse fully, if a positive number the number of levels
+#'   to recurse.
 #' @inheritParams path_filter
 #' @param all If `TRUE` hidden files are also returned.
 #' @param fail Should the call fail (the default) or warn if a file cannot be
@@ -28,7 +29,7 @@
 #' # Create a shorter link
 #' link_create(system.file(package = "base"), "base")
 #'
-#' dir_ls("base", recursive = TRUE, glob = "*.R")
+#' dir_ls("base", recurse = TRUE, glob = "*.R")
 #'
 #' dir_map("base", identity)
 #'
@@ -39,13 +40,13 @@
 #' # Cleanup
 #' link_delete("base")
 #' \dontshow{setwd(.old_wd)}
-dir_ls <- function(path = ".", all = FALSE, recursive = FALSE, type = "any",
+dir_ls <- function(path = ".", all = FALSE, recurse = FALSE, type = "any",
                    glob = NULL, regexp = NULL, invert = FALSE, fail = TRUE, ...) {
   assert_no_missing(path)
 
   old <- path_expand(path)
 
-  files <- as.character(dir_map(old, identity, all, recursive, type, fail))
+  files <- as.character(dir_map(old, identity, all, recurse, type, fail))
 
   path_filter(files, glob, regexp, invert = invert, ...)
 }
@@ -64,33 +65,41 @@ directory_entry_types <- c(
 #' @rdname dir_ls
 #' @param fun A function, taking one parameter, the current path entry.
 #' @export
-dir_map <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any", fail = TRUE) {
+dir_map <- function(path = ".", fun, all = FALSE, recurse = FALSE, type = "any", fail = TRUE) {
   assert_no_missing(path)
+
+  if (is.logical(recurse)) {
+    if (isTRUE(recurse)) {
+      recurse <- -1
+    } else {
+      recurse <- 0
+    }
+  }
 
   type <- match.arg(type, names(directory_entry_types), several.ok = TRUE)
 
   old <- path_expand(path)
 
-  dir_map_(old, fun, all, sum(directory_entry_types[type]), recursive, fail)
+  dir_map_(old, fun, all, sum(directory_entry_types[type]), recurse, fail)
 }
 
 #' @rdname dir_ls
 #' @export
-dir_walk <- function(path = ".", fun, all = FALSE, recursive = FALSE, type = "any", fail = TRUE) {
+dir_walk <- function(path = ".", fun, all = FALSE, recurse = FALSE, type = "any", fail = TRUE) {
   assert_no_missing(path)
 
   old <- path_expand(path)
 
-  dir_map(old, fun, all, recursive, type, fail)
+  dir_map(old, fun, all, recurse, type, fail)
   invisible(path_tidy(path))
 }
 
 #' @rdname dir_ls
 #' @export
-dir_info <- function(path = ".", all = FALSE, recursive = FALSE,
+dir_info <- function(path = ".", all = FALSE, recurse = FALSE,
                      type = "any", regexp = NULL, glob = NULL, fail = TRUE, ...) {
   assert_no_missing(path)
 
-  file_info(dir_ls(path = path, all = all, recursive = recursive, type = type,
+  file_info(dir_ls(path = path, all = all, recurse = recurse, type = type,
     regexp = regexp, glob = glob, fail = fail, ...), fail = fail)
 }