+++ R/tree.R
@@ -5,8 +5,8 @@
 #' @param ... Additional arguments passed to [dir_ls].
 #'
 #' @export
-dir_tree <- function(path = ".", recursive = TRUE, ...) {
-  files <- dir_ls(path, recursive = recursive, ...)
+dir_tree <- function(path = ".", recurse = TRUE, ...) {
+  files <- dir_ls(path, recurse = recurse, ...)
   by_dir <- split(files, path_dir(files))
 
   ch <- box_chars()